Started to implement the method acceptRequest()
Started to implement the method acceptRequest(). This method is used to accept a loan request and to transfer the requested amount.

This method is a preliminary version only. In particular, we do not currently add the transfer to the transactions database.
Also, there is currently no way to save the loan and to remind the requestor to repay.